// src/app/models/user.ts
export class User {
    id_type: string;
    id_num: string;
    id_link: string;
    user: string;
    first_Name: string;
    middle_Name:string;
    last_Name: string;
    emailId: string;
    password:string;
    mobile:string;
    state : string;
    city:string;
    pincode: string;
    token: string;
}